Refresh Token Finished
We now issue refresh tokens, and are able to re-issue new JWTs when a JWT expires.

Refresh tokens are saved into DB as a UUID - this may not be the best approach and should technically encrypt (hash) the tokens with a salt such that if the DB gets hacked no one can actually use the cookies but for now it should be fine without any hashing.