Fixed post-merge bugs
most of them were re-naming var. most notable issue was Hybernate creating a new instance of the User object I just saved, and setting it as the primary instance managed by the session - in turn i was pulling my hair beacause I was trying to save the older instnace, when a new instance was now created, and hybernate saw two instance with the same primary key in session!