Fixed Sliding window bug
Very silly bug: the sorted set is afterall a set, and does not allow duplicate values, thus we cannot give value of "" to all entries despite giving a different score. This results in the set ever having only one item!

Fixed the bug and rate limiter seems to work; will make a python script to stress test it and perhaps move the Rate limiter an aspect so we can pick and choose which endpoints to rate limit.