Pipelined + Tested Rate limiters w/ concurent users
Pipelined some commands so that we do not send some commands one by one to Redis. This is an attempt to microptimize and reduce RTT for Redis interactions.

Stress tested the Rate limiter with a Python Script, by sending multiple request from different threads to a toy API in Spring. I used the x-forwarded-for header so that Spring reads unique IPs on each request.

When we deploy the website we can re-test the rate limiter with proxies, I have the script written for that (it's the same one I used now, but will need a tiny bit of tweaking).